pyproject.toml
[project]
name = "Portal Tabor"
version = "0.1.5"
description = ""
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "flet==0.26.0",
  "python-multipart==0.0.12",
  "pandas==2.2.3",
  "google-auth == 2.18.0",
  "google-api-python-client == 2.88.0"
]

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "tabor.cat"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "Portal Tabor"

# company name to display in about app dialogs
company = "Tabor"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2024 by Escola Tabor"

[tool.flet.app]
path = "src"